#49 ジャンルの管理方法をストレージに変更
