+++ client/src/App.tsx
@@ -14,7 +14,6 @@ import AdminMusic from './components/admin/musics';
 import EditMusicScore from './components/score/edit';
 import PrivateRoute from './lib/PrivateRoute';
 import AuthRoute from './lib/AuthRoute';
-import { GenreProvider } from './components/store/useGenre';
 
 const queryClient = new QueryClient();
 
@@ -22,29 +21,27 @@ function App() {
   return (
     <>
       <QueryClientProvider client={queryClient}>
-        <GenreProvider>
-          <Router>
-            <Routes>
-              <Route element={<PrivateRoute />}>
-                <Route path='/my-list/' element={<MyList />} />
-                <Route path='/music/' element={<MusicList />} />
-                <Route path='/scores/' element={<ScoresList />} />
-                <Route path='/register-music-score/' element={<RegisterMusicScore />} />
-                <Route path='/register-music-score/:musicId' element={<RegisterMusicScore />} />
-                <Route path='/register-music-score/:musicId/:musicDifficulty' element={<RegisterMusicScore />} />
-                <Route path='/edit-score/:scoreId' element={<EditMusicScore />} />
-                <Route path='/score/:musicId/:musicDifficulty' element={<MusicScoreList />} />
-                <Route path='/my-list/:myListId' element={<MyListDetail />} />
-              </Route>
-              <Route element={<AuthRoute />}>
-                <Route path='/admin' element={<AdminTop />} />
-                <Route path='/admin/users' element={<AdminUser />} />
-                <Route path='/admin/musics' element={<AdminMusic />} />
-              </Route>
-              <Route path='/' element={<Home />} />
-            </Routes>
-          </Router>
-        </GenreProvider>
+        <Router>
+          <Routes>
+            <Route element={<PrivateRoute />}>
+              <Route path='/my-list/' element={<MyList />} />
+              <Route path='/music/' element={<MusicList />} />
+              <Route path='/scores/' element={<ScoresList />} />
+              <Route path='/register-music-score/' element={<RegisterMusicScore />} />
+              <Route path='/register-music-score/:musicId' element={<RegisterMusicScore />} />
+              <Route path='/register-music-score/:musicId/:musicDifficulty' element={<RegisterMusicScore />} />
+              <Route path='/edit-score/:scoreId' element={<EditMusicScore />} />
+              <Route path='/score/:musicId/:musicDifficulty' element={<MusicScoreList />} />
+              <Route path='/my-list/:myListId' element={<MyListDetail />} />
+            </Route>
+            <Route element={<AuthRoute />}>
+              <Route path='/admin' element={<AdminTop />} />
+              <Route path='/admin/users' element={<AdminUser />} />
+              <Route path='/admin/musics' element={<AdminMusic />} />
+            </Route>
+            <Route path='/' element={<Home />} />
+          </Routes>
+        </Router>
       </QueryClientProvider>
     </>
   );
